+++ packages/preview-middleware-client/src/adp/quick-actions/common/create-page-action.ts
@@ -1,7 +1,6 @@
 import OverlayRegistry from 'sap/ui/dt/OverlayRegistry';
 import FlexCommand from 'sap/ui/rta/command/FlexCommand';
 
-import { FeatureService } from '../../../cpe/feature-service';
 import { DialogNames, handler } from '../../init-dialogs';
 import { QuickActionContext, SimpleQuickActionDefinition } from '../../../cpe/quick-actions/quick-action-definition';
 import { SimpleQuickActionDefinitionBase } from '../simple-quick-action-base';
@@ -22,10 +21,7 @@ export class AddPageActionQuickAction extends SimpleQuickActionDefinitionBase im
     async initialize(): Promise<void> {
         const appType = getApplicationType(this.context.rta.getRootControlInstance().getManifest());
         const version = await getUi5Version();
-        if (
-            FeatureService.isFeatureEnabled('cpe.beta.quick-actions') === false ||
-            (appType === 'fe-v4' && isLowerThanMinimalUi5Version(version, { major: 1, minor: 130 }))
-        ) {
+        if (appType === 'fe-v4' && isLowerThanMinimalUi5Version(version, { major: 1, minor: 130 })) {
             return;
         }
         return super.initialize();