+++ packages/preview-middleware-client/src/adp/quick-actions/fe-v2/registry.ts
@@ -43,14 +43,14 @@ export default class FEV2QuickActionRegistry extends QuickActionDefinitionRegist
                 definitionGroups.push({
                     title: 'LIST REPORT',
                     definitions: [
-                        ToggleClearFilterBarQuickAction,
                         AddControllerToPageQuickAction,
+                        AddPageActionQuickAction,
+                        ToggleClearFilterBarQuickAction,
+                        ToggleSemanticDateRangeFilterBar,
                         ChangeTableColumnsQuickAction,
                         AddTableActionQuickAction,
                         AddTableCustomColumnQuickAction,
-                        AddPageActionQuickAction,
-                        EnableTableFilteringQuickAction,
-                        ToggleSemanticDateRangeFilterBar
+                        EnableTableFilteringQuickAction
                     ],
                     view,
                     key: name + index
@@ -60,9 +60,10 @@ export default class FEV2QuickActionRegistry extends QuickActionDefinitionRegist
                     title: 'OBJECT PAGE',
                     definitions: [
                         AddControllerToPageQuickAction,
-                        ChangeTableColumnsQuickAction,
+                        AddPageActionQuickAction,
                         AddHeaderFieldQuickAction,
                         AddCustomSectionQuickAction,
+                        ChangeTableColumnsQuickAction,
                         AddTableActionQuickAction,
                         AddTableCustomColumnQuickAction
                     ],