+++ packages/preview-middleware-client/src/adp/quick-actions/fe-v4/registry.ts
@@ -38,14 +38,14 @@ export default class FEV4QuickActionRegistry extends QuickActionDefinitionRegist
                 definitionGroups.push({
                     title: 'LIST REPORT',
                     definitions: [
-                        ToggleClearFilterBarQuickAction,
                         AddControllerToPageQuickAction,
-                        ChangeTableColumnsQuickAction,
                         AddPageActionQuickAction,
+                        ToggleClearFilterBarQuickAction,
+                        ToggleSemanticDateRangeFilterBar,
+                        ChangeTableColumnsQuickAction,
                         AddTableActionQuickAction,
                         AddTableCustomColumnQuickAction,
                         EnableTableFilteringQuickAction,
-                        ToggleSemanticDateRangeFilterBar
                     ],
                     view,
                     key: name + index
@@ -55,10 +55,10 @@ export default class FEV4QuickActionRegistry extends QuickActionDefinitionRegist
                     title: 'OBJECT PAGE',
                     definitions: [
                         AddControllerToPageQuickAction,
-                        ChangeTableColumnsQuickAction,
+                        AddPageActionQuickAction,
                         AddHeaderFieldQuickAction,
                         AddCustomSectionQuickAction,
-                        AddPageActionQuickAction,
+                        ChangeTableColumnsQuickAction,
                         AddTableActionQuickAction,
                         AddTableCustomColumnQuickAction
                     ],